Use marshal serialization shim
This change modifies the manner in which we serialize offload requests.
We use 'marshal' over 'cPickle' and insert a shim to the uwsgin python
plugin mule msg hook and tentatively process messages. If the message
fails to deserialize with 'marshal', we know its a uwsgidecorators message
and pass it along to the appropriate dispatcher