Do not necessarily raise an exception when yelp_clog.json is not found
This is a more backwards-compatible approach when using
`clog.readers.find_tail_host`.
If an old user of yelp-clog did not have a /etc/yelp_clog.json file,
the behaviour will be the same as it used to be.

An Exception can be raised using a flag so future implementations can
choose to handle the Exception or not.