Do not raise exception when host mapping file is missing
This is a more backwards-compatible approach when using
`clog.readers.find_tail_host`. If an old code using yelp-clog
did not have a /etc/yelp_clog.json file, the behaviour will be
the same as it used to (silently fail).

I believe the motivation behind raising any `IOError` was due
to some packages/services silently failing when instantiating
a `StreamTailer` and /etc/yelp_clog.json was not present, hence
the logging in the constructor.