Default to find_tail_host when no host/port is given in StreamTailer
If no host/port is given for a StreamTailer, it used to look at the
configuration in the `scribe_tail_services` list. It would fail if
either this list is empty or the configuration cannot be found.

With this change applied, if no configuration can be found, the host
will be discovered using `find_tail_host`. The port can be discovered
thanks to a new configuration parameter (defaults to port 3535).