Add memory buffering logic to MonkLogger
If the producer fails (e.g: Monk is restarting and is down for a while),
start buffering log lines in memory, up to a certain size threshold
(older messages will be dropped if the size of the buffer grows too much).

Buffering will stop after a configurable amount of time, all buffered
log lines will be flushed back into Monk (this would slow down the
.log_line call when flushing happens, but it should not be too much of a
trouble). If the producer fails again during flushing, all lines will be
added back to the buffer.

This does not handle STREAM_NOT_FOUND use case (it will start
buffering), but this will be tackled later.